hamadi

+++ database/migrations/2023_09_03_054337_create_org_types_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->unsignedInteger('price');
-            $table->unsignedBigInteger('office_id');
+            $table->unsignedInteger('office_id');
             $table->foreign('office_id')->references('id')->on('offices');
             $table->timestamps();
         });
